Import, read desc if you want to know what it is
This imports the project I call dictDialogue.. This is basically a bunch of dialogue generating functions and npc movement ai and also some radiant ai type context awareness, it also does path following and a couple other things. Npc leaders that get followed and such. It's a general test project to make npcs in main game but will surely be used as a basis for many other projects in the future.