Set up steering behavior things(very rudimentary)
There are now npcs and nodes that give and take desire/repellant signals and whatnot. It's pretty janky but other than the stuff on the beach it's the only real project I have related to this, some smoothing, adjustments, and more moving pieces should make it look a little nicer. Anyway.