Imported crystal lights and exported steering npcs
This brings in the crystal lighting engine because it's sick as hell and i need somewhere to test it but also takes the steering behavior npcs I was setting up and moved them to main project because I realized that the debug manager is hyper useful for debugging moving, effect variant things like npcs and influence nodes. So with the console i can drag them around, watch their desire cones, change values on the fly, spawn, delete, ect. It's just a much, much better environment for testing and debugging those so I moved them to main game as guests. But yeah, the lighting engine is incredible, and complex so i'll have to spend a while figuring out what half of it does. That is all.