Yarn dialogues, speech bubbles for multichoice (eh
The black dot in the game can now output dialogue bubbles to tell the story of the yarn dialogue plus has the option to select response options in a way and show them, it's not great, but not bad either. It ends and restarts with a button too. Should be pretty solid. And having yarn working is so cool, since it's just a text editor with some built in tricks vs hacking in those stupid dialogue chains with the FC system... Cool!