Remade speech bubble sizing + mouse choices
The thing is, one, using the size of bubble that fits the goal is the best case scenario, it's as small as it can be at al times, but also, the shader math is somewhat simpler when you're back to knowing that the size of the bubble is 1,1 and stuff like that. So I redid that, made it look the same as before-ish but with nicer mechanics. I also updated the spacings on multi choice and text buffers. You can now simply draw your mouse over the text to highlight the option you want.. Pretty neat. Just kinda messing around, mostly