I fixed it by changing like 3 things
There's a few improvements to readability, a little cross and arrow indicator. There's a way to set the direction of the options as an array and have it auto draw and detect them at those positions if you want 2, 4, 10 options.. It can be skipped and fade and stuff. So now I guess it's what I wanted, literally this morning I felt like it was terrible and now I like it. That's how she goes innit. Now to expand it and add some variable conditions and more than 2 dialogues to play for different characters and we chillin, idk when I'll move it to main game, probably when I get the bubble shaders figured out and a few options for it. Cheers