Repeat option culling, value dictionaries
The dialogue that you get is now based on a dictionary of sorts that stores a bunch of game values like whether you've been told a secret or done a quest or talked to some a certain number of times. It's really simple to create the branching ifs that lead to different responses (mostly at the beginning that lead to branches within the dialogue) and editing multiple values at once is easy too, you can set them in the dialogue and then check for them wherever. Then when you talk to someone you load in the dialogue that fits with options for going to other places and you're chilling. I also added a system for either culling options that you've already seen or simply fading them but still having them clickable. This is a toggle via a macro but I suppose it could be per option or per dialogue as well. It's quite snazzy though. The more complex dialogue frame, conditional responses and comments, and control over showing and not showing certain responses or options makes this pretty close to done. I still want to be able to add and remove options to multi choices based on criteria which is also pretty easy but not yet done.