Fixed optional reactions and added emote marker
There were some "quirks" with the last build but now the responses marked with #react are skipped nicely and don't show lines after the first, which also means you can do multi line reactions which is sick. As long as they eventually drop back to the first it's chillin. But I also made a little exclamation emote that looks pretty nice. I think that idea should definitely be expanded because it's nice to be able to mark certain things in the game, exclamations for npc ideas, question marks for ai... I dunno, maybe that's too silly. But for this I like it, it shows, and maybe it's the player mark too, since they can do these reactions too. But yeah, nice little mark emotes. And they indicate that there is an option for reaction, though it should probably be a little more subtle...