Non recreated bubbles and smooth resizing
You can now use a metadata tag to trigger bubble recreations (or leave them out to have the dialogue clear and retype in the same bubble, which feels much chiller, but reforming is useful for certain things too) and also when you're reforming bubbles the bubbles now approach the goal values of the new bubble smoothly and thus shift and grow/shrink at a nice speed instead of jumping straight to the new size. This of course is only needed because bubbles can be refitted with multiple lines but it's quite nice regardless. I think that's most of what I wanted to do with resizing and continued bubbles. The metadata stuff is quite fancy and gives nice control over when bubbles do what, which has artistic value.. So yeah, with this I'm not sure what else to add. THough i'm sure I'll think of something! Also, the code and function structure is a bit of a mess and should be cleaned up... But anyway. Nice.