Cleaned lag room lag, dialogue script stuff
I'm on my laptop so i cleared some random stuff that was slowing it down still, but other than that I expanded the npc personality response dialogues. Honestly, it's tough, complex.. but when I work on this i get hyped as hell man. It's so cool, I feel like with the right expanding and maybe a rework or two to the structure once I have more practice with this kind of thing I could make some insanely cool things. Between the emotion reactions, behavior perception, and dialogue dictionaries there's so much cool dynamic stuff here. So much! Like, all the stuff that makes skyrim dynamic and fun (or oblivion really..) is here and more, I could do some crazy interactions and mixing fragments and interweaving different responses lets me branch things so deeply with not too much work. It's super super cool... But i feel like i never work on it, it's hard to get in to. It's vague and esoteric work but if you just plod forward into the unknown it's really not hard to work on. At least until you screw it up but then you know what to improve and that's good too. Point being, I really want to work more on this, like a lot more, but I haven't. And that makes me sad! I should... You'll see if I do.