Formatting adjustments, give items to npcs
Now most (i barely checked i probably missed a bunch) dialogue scripts take a subject (the one doing the response) and "otherSpeaker" aka the conversation partner, the one doing the talking or thinking is the subject so in a response for age the person who is getting their age is the subject. You understand, anyway, the other thing I did was make it so that you can "pick up" items (in the sense of space game holding it over your head) and then give them to npcs by asking if they want it, this then makes them judge it and respond in a simple-ish way, based on their opinion of the item. This is cool but needs to be way more complex and also starts to rub on the ideas that the npc personality would have an emotion set with it, like the values I have now are personality traits but they aught to have emotions in the moment that are affected (also relationship values, though this is more complex..?) by what you do and say with them. For example if you give them an item they hate their opinion of you will go down and fear go up and whatever, if you give them items they like it goes up. This is simple game stuff that you see everywhere. However there could be other facets of this, like if you give an npc a super powerful terrible artifact they don't just get unhappy they get scared and run away. Giving items is a simple ish case but perhaps there could be humor and other such things attached. This would also apply to conversation, your dialogue interactions would carry emotional values that would get filtered through their personality and emotional alchemy systems to produce responses. The emotional alchemy is what I call the system where emotions mix together and give some kind of interesting result, like evil plus humor might not yield laugh plus unhappy, maybe it produces terror or amplifies the humor instead, so some input emotions are completely lost or changed based on some factor here or there. Complex stuff, Idk if i'll be able to do it but i can try. So yeah, I have this outlined more in the docs and comments, there's several layers here, emotional interpretation, dialogue chaining, item and ambient awareness... Other stuff... But anyway, i made the npc pick up the item wowwww... I might be naive.