Emotion updates and interface class, stealing misc
So now player and npcs inherit from the same emotion class, this means that many types of things can play a role in the behavior stuff, which is good. I also updated the stealing logic and info arrows to more properly show and handle stolen items affecting the emotion aspect of npcs.

To go along with that I also made a little toggled display where the debug now shows info, emotions, or both (emotions as colored circles). These emotions also now fade back to .5 (the middle) and are increasingly difficult to force from that center. Basically, the foundations of emotions for npcs. The player can also do ACT functions now, as I mentioned. I also cleaned up and balanced a few things relating to all of this. So, yuh. Pretty neat. It's like a 500 step road where each step feels extremely important, so many pieces that interconnect with each other, each layer is neccessary for the others and yet cannot be made without a decent base around it so I have to update in sheets. It's nice, I love stuff like this from a work flow perspective, just adding and adding and adding in circles until everythings more complex. THen you have to go back and refine down a bunch over the years of use... Haven't gotten to that part though.. Anyway, vibing.