Basic repeated response, comment mood tags, misc
Reworked the inheritance of the difference npcs to let chat npcs have emotions which lets them do some things relating to mood and stuff, though not super well since they can't really respond in any way. Regardless, it's useful and I'll eventually maybe merge the chatter npcs with the others to make some kind of final, all encompassing npc form. Which will be neat.

I also made dialogue tags for jokes, insults, threats, sarcasm, ect so that when you're speaking to an npc certain comments and lines of dialogue in chatterbox will carry emotional changes with them. These are interpreted as archetypes like the high level "sarcasm", "joke", "threat" or manually in the generated responses via influence script calls when choosing the response, so if an npc responds with "how dare you" to a rude question, it's fair to assume that they will get annoyed, distrusting, angry, ect at that comment and so I put the influence there. However it's possible that at some point, some responses are so generated that the resulting comment is hardly understandable to me from the outside. I think it doesn't matter much, (if that day ever comes) I'll make a system for assigning emotionality to responses for that but I doubt I'll have to. For now I have two ways to react to lines given or recieved, for each system respectively. I think it's good but feels a bit clunky tbh.

The last thing I did was add a check in the dict dialogue system for repeated (valid) lines, because I think repeating lines, or maybe even using the same line in a 10 comment range, should probably incite some kind of "I just told you that" kind of response, and so it does now, though only for consecutive comments and it's not exactly sophisticated. But in the future if I have a proper "topics covered" type memory I can check those and return lines based on the impression of those, whether they were important or not giving "How could you forget??" and "well I mentioned..." for example. But that's speculation. That's all for now, i've been writing a ton and like always when one art advances another stagnates... Cheerio bruh.