Switched a bunch of things to inbuilt speed values
Not really an improvement but I'm trying to build the habit of using the built ins because they are actually slightly more rationally set up (that being, stopping movement is less sensible than setting it to 0 (no click on off effect..) and the values are faster to use combined with the convenience of always having access to speed and direction by default. (do you know how many times I calculated direction or speed in main game? Do you?? I can finally be free to just know the speed or direction of the enemy i'm programming on hand.... No more var management to try and prevent lag while using the same vars over and over, copy pasting into each nook... It's not that bad tbh but this is just better in every way so whatever.